+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Changed
 
 - [TD-3569] Only admin and service users can manage templates
+- [TD-4230] Moved templates preprocessing functionality to `td-cache` and allows for multiple domain_ids
 
 ## [4.40.0] 2022-03-14